backend/requirements.txt
# ===== CrewAI Stack =====
# Using latest stable crewai version
crewai>=0.86.0

# ===== LangChain Stack =====
# Compatible with latest crewai
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
groq>=0.11.0
langgraph>=0.2.0

# ===== FastAPI Backend =====
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.6

# ===== Utilities =====
python-dotenv>=1.0.0
requests>=2.32.0
aiofiles>=23.2.0
websockets>=12.0

# ===== Audio =====
SpeechRecognition>=3.10.0
gTTS>=2.5.0
pydub>=0.25.0
pygame>=2.5.0

# ===== Database =====
pymongo>=4.7.0
motor>=3.4.0

# ===== UI & Visualization =====
streamlit>=1.30.0

# ===== System Control (Optional) =====
pycaw>=20240210
comtypes>=1.4.0
psutil>=5.9.0
screen-brightness-control>=0.21.0
